authentication & login resolver

+++ resolvers/notes.js
@@ -20,17 +20,17 @@ const getNotes = async(noteid) =>{
   }
 }
 module.exports = {
-    notes: async () =>{
-        try {
-          const Notes = await Note.find({})
-          return Notes.map(note =>{
-            return {...note._doc, userCreator: getUser(note._doc.userCreator)}
-          })
-        }catch(err){
-          throw err
-        }
-    },
-    note: async ({_id}) =>{
+  notes: async () =>{
+    try {
+      const Notes = await Note.find({})
+      return Notes.map(note =>{
+        return {...note._doc, userCreator: getUser(note._doc.userCreator)}
+      })
+    }catch(err){
+      throw err
+    }
+  },
+  note: async ({_id}) =>{
       try{
         const findNote = await Note.findById(_id)
         return {
@@ -39,37 +39,43 @@ module.exports = {
       }catch(err){
         throw err
       }
-    },
-    deleteNote: async({_id}) =>{
+  },
+  deleteNote: async({_id}, req) =>{
+    if(!req.authUser){
+      throw new Error("Not Authorized")
+    }
         try {
           const note = await Note.findByIdAndRemove({_id: _id})
           return {...note._doc}
         }catch(err){
           throw err
         }
-    },
-    createNote: async ({noteInput}) =>{
-        const note = new Note({
-          title: noteInput.title,
-          content: noteInput.content,
-          image: noteInput.image,
-          userCreator: '5d267e2872bc12546236efd0'
-        })
-        let notes;
-        try {
-          // saving and return notes
-          const result = await note.save()
-          notes = {
-            ...result._doc,
-            userCreator: getUser(result._doc.userCreator)
-          }
-          const findUser = await User.findById('5d267e2872bc12546236efd0')
-          if(!findUser) return new Error('User not found')
-          findUser.createdNotes.push(note)
-          await findUser.save()
-          return notes
-        }catch (err){
-          throw err
-        }
+  },
+  createNote: async ({noteInput}, req) =>{
+    if(!req.authUser){
+      throw new Error("Not Authorized")
     }
+   const note = new Note({
+     title: noteInput.title,
+     content: noteInput.content,
+     image: noteInput.image,
+     userCreator: req.userId
+   })
+   let notes;
+   try {
+     // saving and return notes
+     const result = await note.save()
+     notes = {
+       ...result._doc,
+       userCreator: getUser(result._doc.userCreator)
+     }
+     const findUser = await User.findById(req.userId)
+     if(!findUser) return new Error('User not found')
+     findUser.createdNotes.push(note)
+     await findUser.save()
+     return notes
+   }catch (err){
+     throw err
+   }
+  }
 }
